requirements.txt
fastapi
uvicorn
langchain
langchain-community
pymupdf
faiss-cpu
sentence-transformers
python-multipart
qdrant-client
requests
google-generativeai
streamlit
